test(asyncapi): use a template
