--- src/rulesets/asyncapi/__tests__/asyncapi-info-contact.ts
@@ -1,46 +1,0 @@
-import type { Spectral } from '../../../spectral';
-import { createWithRules } from './__helpers__/createWithRules';
-import { DiagnosticSeverity } from '@stoplight/types';
-
-const ruleName = 'asyncapi-info-contact';
-
-describe(`Rule '${ruleName}'`, () => {
-  let s: Spectral;
-  let doc: any;
-
-  beforeEach(async () => {
-    s = await createWithRules([ruleName]);
-
-    doc = {
-      asyncapi: '2.0.0',
-      info: {
-        contact: {
-          name: 'stoplight',
-          url: 'stoplight.io',
-          email: 'support@stoplight.io',
-        },
-      },
-    };
-  });
-
-  test('validates a correct object', async () => {
-    const results = await s.run(doc, { ignoreUnknownFormat: false });
-
-    expect(results).toEqual([]);
-  });
-
-  test('return result if contact property is missing', async () => {
-    delete doc.info.contact;
-
-    const results = await s.run(doc, { ignoreUnknownFormat: false });
-
-    expect(results).toEqual([
-      expect.objectContaining({
-        code: ruleName,
-        message: 'Info object should contain `contact` object.',
-        path: ['info'],
-        severity: DiagnosticSeverity.Warning,
-      }),
-    ]);
-  });
-});
