--- src/rulesets/asyncapi/__tests__/asyncapi-info-description.ts
@@ -1,42 +1,0 @@
-import type { Spectral } from '../../../spectral';
-import { createWithRules } from './__helpers__/createWithRules';
-import { DiagnosticSeverity } from '@stoplight/types/';
-
-const ruleName = 'asyncapi-info-description';
-
-describe(`Rule '${ruleName}'`, () => {
-  let s: Spectral;
-  let doc: any;
-
-  beforeEach(async () => {
-    s = await createWithRules([ruleName]);
-
-    doc = {
-      asyncapi: '2.0.0',
-      info: {
-        description: 'Very descriptive list of self explaining consecutive characters.',
-      },
-    };
-  });
-
-  test('validates a correct object', async () => {
-    const results = await s.run(doc, { ignoreUnknownFormat: false });
-
-    expect(results).toEqual([]);
-  });
-
-  test('return result if description property is missing', async () => {
-    delete doc.info.description;
-
-    const results = await s.run(doc, { ignoreUnknownFormat: false });
-
-    expect(results).toEqual([
-      expect.objectContaining({
-        code: ruleName,
-        message: 'AsyncAPI object info `description` must be present and non-empty string.',
-        path: ['info'],
-        severity: DiagnosticSeverity.Warning,
-      }),
-    ]);
-  });
-});
