+++ src/rulesets/asyncapi/__tests__/asyncapi-info-license-url.test.ts
@@ -1,0 +1,34 @@
+import testRule from './__helpers__/tester';
+import { DiagnosticSeverity } from '@stoplight/types/';
+
+testRule('asyncapi-info-license-url', [
+  {
+    name: 'valid case',
+    document: {
+      asyncapi: '2.0.0',
+      info: {
+        license: {
+          url: 'https://github.com/stoplightio/spectral/blob/develop/LICENSE',
+        },
+      },
+    },
+    errors: [],
+  },
+
+  {
+    name: 'url property is missing',
+    document: {
+      asyncapi: '2.0.0',
+      info: {
+        license: {},
+      },
+    },
+    errors: [
+      {
+        message: 'License object should include `url`.',
+        path: ['info', 'license'],
+        severity: DiagnosticSeverity.Warning,
+      },
+    ],
+  },
+]);
