--- src/rulesets/asyncapi/__tests__/asyncapi-info-license-url.ts
@@ -1,44 +1,0 @@
-import { Spectral } from '../../../spectral';
-import { createWithRules } from './__helpers__/createWithRules';
-import { DiagnosticSeverity } from '@stoplight/types';
-
-const ruleName = 'asyncapi-info-license-url';
-
-describe(`Rule '${ruleName}'`, () => {
-  let s: Spectral;
-  let doc: any;
-
-  beforeEach(async () => {
-    s = await createWithRules([ruleName]);
-
-    doc = {
-      asyncapi: '2.0.0',
-      info: {
-        license: {
-          url: 'https://github.com/stoplightio/spectral/blob/develop/LICENSE',
-        },
-      },
-    };
-  });
-
-  test('validates a correct object', async () => {
-    const results = await s.run(doc, { ignoreUnknownFormat: false });
-
-    expect(results).toEqual([]);
-  });
-
-  test('return result if url property is missing', async () => {
-    delete doc.info.license.url;
-
-    const results = await s.run(doc, { ignoreUnknownFormat: false });
-
-    expect(results).toEqual([
-      expect.objectContaining({
-        code: ruleName,
-        message: 'License object should include `url`.',
-        path: ['info', 'license'],
-        severity: DiagnosticSeverity.Warning,
-      }),
-    ]);
-  });
-});
