--- src/rulesets/asyncapi/__tests__/asyncapi-schema.ts
@@ -1,44 +1,0 @@
-import { Spectral } from '../../../spectral';
-import { createWithRules } from './__helpers__/createWithRules';
-import { DiagnosticSeverity } from '@stoplight/types';
-
-const ruleName = 'asyncapi-schema';
-
-describe(`Rule '${ruleName}'`, () => {
-  let s: Spectral;
-  let doc: any;
-
-  beforeEach(async () => {
-    s = await createWithRules([ruleName]);
-
-    doc = {
-      asyncapi: '2.0.0',
-      info: {
-        title: 'Valid AsyncApi document',
-        version: '1.0',
-      },
-      channels: {},
-    };
-  });
-
-  test('validates a correct object', async () => {
-    const results = await s.run(doc, { ignoreUnknownFormat: false });
-
-    expect(results).toEqual([]);
-  });
-
-  test('return result if channels property is missing', async () => {
-    delete doc.channels;
-
-    const results = await s.run(doc, { ignoreUnknownFormat: false });
-
-    expect(results).toEqual([
-      expect.objectContaining({
-        code: ruleName,
-        message: 'Object must have required property `channels`',
-        path: [],
-        severity: DiagnosticSeverity.Error,
-      }),
-    ]);
-  });
-});
