+++ src/rulesets/asyncapi/__tests__/asyncapi-server-no-trailing-slash.test.ts
@@ -1,0 +1,38 @@
+import { DiagnosticSeverity } from '@stoplight/types';
+import testRule from './__helpers__/tester';
+
+testRule('asyncapi-server-no-trailing-slash', [
+  {
+    name: 'valid',
+    document: {
+      asyncapi: '2.0.0',
+      servers: {
+        production: {
+          url: 'stoplight.io',
+          protocol: 'https',
+        },
+      },
+    },
+    errors: [],
+  },
+
+  {
+    name: '{server}.url property ends with a trailing slash',
+    document: {
+      asyncapi: '2.0.0',
+      servers: {
+        production: {
+          url: 'stoplight.io/',
+          protocol: 'https',
+        },
+      },
+    },
+    errors: [
+      {
+        message: 'Server URL should not end with a slash.',
+        path: ['servers', 'production', 'url'],
+        severity: DiagnosticSeverity.Warning,
+      },
+    ],
+  },
+]);
