--- src/rulesets/asyncapi/__tests__/asyncapi-server-no-trailing-slash.ts
@@ -1,45 +1,0 @@
-import type { Spectral } from '../../../spectral';
-import { DiagnosticSeverity } from '@stoplight/types';
-import { createWithRules } from './__helpers__/createWithRules';
-
-const ruleName = 'asyncapi-server-no-trailing-slash';
-
-describe(`Rule '${ruleName}'`, () => {
-  let s: Spectral;
-  let doc: any;
-
-  beforeEach(async () => {
-    s = await createWithRules([ruleName]);
-
-    doc = {
-      asyncapi: '2.0.0',
-      servers: {
-        production: {
-          url: 'stoplight.io',
-          protocol: 'https',
-        },
-      },
-    };
-  });
-
-  test('validates a correct object', async () => {
-    const results = await s.run(doc, { ignoreUnknownFormat: false });
-
-    expect(results).toEqual([]);
-  });
-
-  test('return result if {server}.url property ends with a trailing slash', async () => {
-    doc.servers.production.url = 'stoplight.io/';
-
-    const results = await s.run(doc, { ignoreUnknownFormat: false });
-
-    expect(results).toEqual([
-      expect.objectContaining({
-        code: ruleName,
-        message: 'Server URL should not end with a slash.',
-        path: ['servers', 'production', 'url'],
-        severity: DiagnosticSeverity.Warning,
-      }),
-    ]);
-  });
-});
