--- src/rulesets/asyncapi/__tests__/asyncapi-servers.ts
@@ -1,61 +1,0 @@
-import type { Spectral } from '../../../spectral';
-import { createWithRules } from './__helpers__/createWithRules';
-import { DiagnosticSeverity } from '@stoplight/types';
-
-const ruleName = 'asyncapi-servers';
-
-describe(`Rule '${ruleName}'`, () => {
-  let s: Spectral;
-  let doc: any;
-
-  beforeEach(async () => {
-    s = await createWithRules([ruleName]);
-
-    doc = {
-      asyncapi: '2.0.0',
-      servers: {
-        production: {
-          url: 'stoplight.io',
-          protocol: 'https',
-        },
-      },
-    };
-  });
-
-  test('validates a correct object', async () => {
-    const results = await s.run(doc, { ignoreUnknownFormat: false });
-
-    expect(results).toEqual([]);
-  });
-
-  test('return result if servers property is missing', async () => {
-    delete doc.servers;
-
-    const results = await s.run(doc, { ignoreUnknownFormat: false });
-
-    expect(results).toEqual([
-      expect.objectContaining({
-        code: ruleName,
-        message: 'AsyncAPI object should contain a non empty `servers` object.',
-        path: [],
-        severity: DiagnosticSeverity.Warning,
-      }),
-    ]);
-  });
-
-  test('return result if servers property is empty', async () => {
-    delete doc.servers.production;
-    expect(doc.servers).toEqual({});
-
-    const results = await s.run(doc, { ignoreUnknownFormat: false });
-
-    expect(results).toEqual([
-      expect.objectContaining({
-        code: ruleName,
-        message: 'AsyncAPI object should contain a non empty `servers` object.',
-        path: ['servers'],
-        severity: DiagnosticSeverity.Warning,
-      }),
-    ]);
-  });
-});
