--- src/rulesets/asyncapi/__tests__/asyncapi-tag-description.ts
@@ -1,45 +1,0 @@
-import type { Spectral } from '../../../spectral';
-import { createWithRules } from './__helpers__/createWithRules';
-import { DiagnosticSeverity } from '@stoplight/types';
-
-const ruleName = 'asyncapi-tag-description';
-
-describe(`Rule '${ruleName}'`, () => {
-  let s: Spectral;
-  let doc: any;
-
-  beforeEach(async () => {
-    s = await createWithRules([ruleName]);
-
-    doc = {
-      asyncapi: '2.0.0',
-      tags: [
-        {
-          name: 'a tag',
-          description: "I'm a tag.",
-        },
-      ],
-    };
-  });
-
-  test('validates a correct object', async () => {
-    const results = await s.run(doc, { ignoreUnknownFormat: false });
-
-    expect(results).toEqual([]);
-  });
-
-  test('return result if description property is missing', async () => {
-    delete doc.tags[0].description;
-
-    const results = await s.run(doc, { ignoreUnknownFormat: false });
-
-    expect(results).toEqual([
-      expect.objectContaining({
-        code: ruleName,
-        message: 'Tag object should have a `description`.',
-        path: ['tags', '0'],
-        severity: DiagnosticSeverity.Warning,
-      }),
-    ]);
-  });
-});
